Put mcp tool stuff into modals-within-modals
